feat: add new models for handling projects creation
