+++ src/main/groovy/io/saagie/plugin/dataops/models/Project.groovy
@@ -8,6 +8,9 @@ class Project implements IMapable {
     Integer jobsCount
     String status
 
+    List<TechnologyByCategory> technologyByCategory = []
+    List<SecurityGroup> authorizedGroups = []
+
     @Override
     Map toMap() {
         if (id && name && creator && description && jobsCount && status) {
@@ -18,6 +21,13 @@ class Project implements IMapable {
                 description: description,
                 jobsCount  : jobsCount,
                 status     : status
+            ]
+        } else if (!id && name) {
+            return [
+                name: name,
+                description: description,
+                technologiesByCategory: technologyByCategory ? technologyByCategory.collect({ it.toMap() }) : null,
+                authorizedGroups: authorizedGroups.collect({ it.toMap() })
             ]
         }
         return null
