+++ src/test/groovy/io/saagie/plugin/models/TechnologyByCategoryTest.groovy
@@ -1,0 +1,66 @@
+package io.saagie.plugin.models
+
+import io.saagie.plugin.dataops.models.TechnologyByCategory
+import spock.lang.Specification
+import spock.lang.Title
+
+@Title("Tests on TechnologyByCategory model")
+class TechnologyByCategoryTest extends Specification {
+
+    def "the model should have default values"() {
+        given:
+        TechnologyByCategory technologyByCategory = new TechnologyByCategory()
+
+        expect:
+        !technologyByCategory.exists()
+        !technologyByCategory.toMap()
+    }
+
+    def "the model should return a map with only required provided data"() {
+        given:
+        TechnologyByCategory technologyByCategory = new TechnologyByCategory()
+        technologyByCategory.jobCategory = "job-category"
+
+        expect:
+        technologyByCategory.exists()
+        technologyByCategory.jobCategory == "job-category"
+    }
+
+    def "the model should return a correct map when the toMap method is called"() {
+        given:
+        TechnologyByCategory technologyByCategory = new TechnologyByCategory()
+        technologyByCategory.jobCategory = "job-category"
+
+        when:
+        Map result = technologyByCategory.toMap()
+
+        then:
+        technologyByCategory.exists()
+        result.size() == 2
+        result.containsKey("technologies")
+        result.containsKey("jobCategory")
+        result.jobCategory == "job-category"
+        result.technologies == null
+    }
+
+    def "the model should return a correct map when the toMap method is called and technologies are provided"() {
+        given:
+        TechnologyByCategory technologyByCategory = new TechnologyByCategory()
+        technologyByCategory.jobCategory = "job-category"
+        technologyByCategory.technologyid = [
+            "techno1", "techno2"
+        ]
+
+        when:
+        Map result = technologyByCategory.toMap()
+
+        then:
+        technologyByCategory.exists()
+        result.size() == 2
+        result.containsKey("technologies")
+        result.containsKey("jobCategory")
+        result.jobCategory == "job-category"
+        result.technologies == [[id: "techno1"], [id: "techno2"]]
+    }
+
+}
