+++ src/test/groovy/io/saagie/plugin/tasks/job/JobRunTaskTests.groovy
@@ -1,9 +1,10 @@
-package io.saagie.plugin.technology
+package io.saagie.plugin.tasks.job
 
 import okhttp3.mockwebserver.MockResponse
 import okhttp3.mockwebserver.MockWebServer
 import org.gradle.testkit.runner.BuildResult
 import org.gradle.testkit.runner.GradleRunner
+import org.gradle.testkit.runner.UnexpectedBuildFailure
 import org.junit.Rule
 import org.junit.rules.TemporaryFolder
 import spock.lang.Shared
@@ -12,11 +13,11 @@ import spock.lang.Title
 
 import static org.gradle.testkit.runner.TaskOutcome.FAILED
 
-@Title("projectsListTechnologies task tests")
-class TechnologyTests extends Specification {
+@Title('projectsRunJob task tests')
+class JobRunTaskTests extends Specification {
     @Rule TemporaryFolder testProjectDir = new TemporaryFolder()
     @Shared MockWebServer mockWebServer = new MockWebServer()
-    @Shared String taskName = 'projectsListTechnologies'
+    @Shared String taskName = 'projectsRunJob'
 
     File buildFile
     File jobFile
@@ -55,11 +56,13 @@ class TechnologyTests extends Specification {
         gradle(true, arguments)
     }
 
-    def "projectsListTechnologies task should list technologies of a project"() {
+    def "projectsRunJob should run a job and return the job instance id and status"() {
         given:
-        def mockedResponse = new MockResponse()
-        mockedResponse.responseCode = 200
-        mockedResponse.body = """{"data":{"technologies":[{"id":"c3cadcad-fjrehf-4f7d-a3a5-frefer","label":"Spark","isAvailable":true,"icon":"spark","features":[]},{"id":"freojfier-c18b-4ecd-b61f-fjerijfiej","label":"Python","isAvailable":true,"icon":"python","features":[]},{"id":"fkiorjeiofer-c18b-4ecd-b61f-jkfijorjferferf","label":"Python","isAvailable":true,"icon":"python","features":[]},{"id":"frefreferfe-26bd-4f7d-a3a5-frejferiuh","label":"Spark","isAvailable":true,"icon":"spark","features":[]}]}}"""
+        def mockedRunJobResponse = new MockResponse()
+        mockedRunJobResponse.responseCode = 200
+        mockedRunJobResponse.body = '''{"data":{"runJob":{"id":"job-instance-id","status":"REQUESTED"}}}'''
+        mockWebServer.enqueue(mockedRunJobResponse)
+
         buildFile << """
             saagie {
                 server {
@@ -69,27 +72,51 @@ class TechnologyTests extends Specification {
                     environment = 2
                 }
 
-                project {
-                    id = 'dezdezjiodjei-892a-2342-8552-5be4b6de5df4'
+                job {
+                    id = "jobId"
                 }
             }
         """
-        mockWebServer.enqueue(mockedResponse)
 
         when:
-        def result = gradle(taskName)
+        BuildResult result = gradle(taskName)
 
         then:
         !result.output.contains('"data"')
-        result.output.contains('"label"')
-        result.output.contains('"features"')
+        result.output.contains('"id"')
+        result.output.contains('"status"')
     }
 
-    def "listed technologies should not show duplicated entries"() {
+    def "projectsRunJob should fail if job or job id is null"() {
         given:
-        def mockedResponse = new MockResponse()
-        mockedResponse.responseCode = 200
-        mockedResponse.body = """{"data":{"technologies":[{"id":"python-id","label":"Python","isAvailable":true,"icon":"python","features":[]},{"id":"bash-id","label":"Bash","isAvailable":true,"icon":"bash","features":[]},{"id":"bash-id","label":"Bash","isAvailable":true,"icon":"bash","features":[]}]}}"""
+        buildFile << """
+            saagie {
+                server {
+                    url = 'http://localhost:9000'
+                    login = 'fake.user'
+                    password = 'ThisPasswordIsWrong'
+                    environment = 2
+                }
+            }
+        """
+
+        when:
+        BuildResult result = gradle(taskName)
+
+        then:
+        UnexpectedBuildFailure e = thrown()
+        result == null
+        e.message.contains("Missing params in plugin configuration: https://github.com/saagie/gradle-saagie-dataops-plugin/wiki/${taskName}")
+        e.getBuildResult().task(":${taskName}").outcome == FAILED
+    }
+
+    def "projectsRunJob should fail if job id doesn't exists"() {
+        given:
+        def mockedRunJobResponse = new MockResponse()
+        mockedRunJobResponse.responseCode = 200
+        mockedRunJobResponse.body = '''{"data":null,"errors":[{"message":"Unexpected error","extensions":null,"path":null}]}'''
+        mockWebServer.enqueue(mockedRunJobResponse)
+
         buildFile << """
             saagie {
                 server {
@@ -99,18 +126,19 @@ class TechnologyTests extends Specification {
                     environment = 2
                 }
 
-                project {
-                    id = 'dezdezjiodjei-892a-2342-8552-5be4b6de5df4'
+                job {
+                    id = 'bad-id'
                 }
             }
         """
-        mockWebServer.enqueue(mockedResponse)
 
         when:
-        def result = gradle(taskName)
+        BuildResult result = gradle(taskName)
 
         then:
-        result.output.count('bash-id') == 1
-        result.output.count('python-id') == 1
+        UnexpectedBuildFailure e = thrown()
+        result == null
+        e.message.contains('Something went wrong when requesting the job to run: {"data":null,"errors":[{"message":"Unexpected error","extensions":null,"path":null}]}')
+        e.getBuildResult().task(":${taskName}").outcome == FAILED
     }
 }
