+++ src/test/groovy/io/saagie/plugin/tasks/pipeline/PipelineDeleteTaskTests.groovy
@@ -1,4 +1,4 @@
-package io.saagie.plugin.job
+package io.saagie.plugin.tasks.pipeline
 
 import okhttp3.mockwebserver.MockResponse
 import okhttp3.mockwebserver.MockWebServer
@@ -13,11 +13,11 @@ import spock.lang.Title
 
 import static org.gradle.testkit.runner.TaskOutcome.FAILED
 
-@Title('projectsListJobs task tests')
-class JobListTaskTests extends Specification {
+@Title('projectsDeletePipeline task tests')
+class PipelineDeleteTaskTests extends Specification {
     @Rule TemporaryFolder testProjectDir = new TemporaryFolder()
     @Shared MockWebServer mockWebServer = new MockWebServer()
-    @Shared String taskName = 'projectsListJobs'
+    @Shared String taskName = 'projectsDeletePipeline'
 
     File buildFile
     File jobFile
@@ -56,50 +56,46 @@ class JobListTaskTests extends Specification {
         gradle(true, arguments)
     }
 
-    def "projectsListJobs task should list jobs on a project"() {
+    def "projectsDeletePipeline should delete a pipeline the deletion status"() {
         given:
-        def mockedResponse = new MockResponse()
-        mockedResponse.responseCode = 200
-        mockedResponse.body = """{"data":{"jobs":[{"name":"test2","description":"","countJobInstance":1,"versions":[{"number":1}],"category":"Processing","technology":{"id":"frefref-c18b-4ecd-b61f-frefefreff","label":"Python","isAvailable":true},"isScheduled":false,"cronScheduling":null,"scheduleStatus":null,"alerting":null,"isStreaming":false,"creationDate":"2019-03-15T14:06:49.053Z","migrationStatus":null,"migrationProjectId":null,"isDeletable":true},{"name":"test 2","description":"","countJobInstance":4,"versions":[{"number":2},{"number":0}],"category":"Processing","technology":{"id":"dezded-26bd-4f7d-a3a5-dezdedzdz","label":"Spark","isAvailable":true},"isScheduled":false,"cronScheduling":null,"scheduleStatus":null,"alerting":null,"isStreaming":false,"creationDate":"2019-03-11T09:32:46.424Z","migrationStatus":null,"migrationProjectId":null,"isDeletable":true}]}}"""
-        buildFile << '''
+        def mockedDeletePipelineResponse = new MockResponse()
+        mockedDeletePipelineResponse.responseCode = 200
+        mockedDeletePipelineResponse.body = '''{"data":{"deletePipeline":true}}'''
+        mockWebServer.enqueue(mockedDeletePipelineResponse)
+
+        buildFile << """
             saagie {
                 server {
                     url = 'http://localhost:9000'
-                    login = 'fake.user'
-                    password = 'ThisPasswordIsWrong'
+                    login = 'test.user'
+                    password = 'password'
                     environment = 2
                 }
 
-                project {
-                    id = 'projectId'
+                pipeline {
+                    id = "pipeline-id"
                 }
             }
-        '''
-        mockWebServer.enqueue(mockedResponse)
+        """
 
         when:
         BuildResult result = gradle(taskName)
 
         then:
+        notThrown(Exception)
         !result.output.contains('"data"')
-        result.output.contains('"name"')
-        result.output.contains('"id"')
-        result.output.contains('"countJobInstance"')
+        result.output.contains('{"status":"success"}')
     }
 
-    def "projectsListJobs task should fail if bad project config is provided"() {
+    def "projectsDeletePipeline should fail if no pipeline id is provided"() {
         given:
         buildFile << """
             saagie {
                 server {
                     url = 'http://localhost:9000'
-                    login = 'fake.user'
-                    password = 'ThisPasswordIsWrong'
+                    login = 'test.user'
+                    password = 'password'
                     environment = 2
-                }
-
-                project {
-
                 }
             }
         """
@@ -114,27 +110,27 @@ class JobListTaskTests extends Specification {
         e.getBuildResult().task(":${taskName}").outcome == FAILED
     }
 
-    def "projectsListJobs task should fail if a wrong project id is provided"() {
+    def "projectsDeletePipeline should fail if pipeline id doesn't exists"() {
         given:
-        def mockedResponse = new MockResponse()
-        mockedResponse.responseCode = 200
-        mockedResponse.body = """{"data":null,"errors":[{"message":"Unexpected error","extensions":null,"path":null}]}"""
+        def mockedDeletePipelineResponse = new MockResponse()
+        mockedDeletePipelineResponse.responseCode = 200
+        mockedDeletePipelineResponse.body = '''{"data":null,"errors":[{"message":"Unexpected error","extensions":null,"path":null}]}'''
+        mockWebServer.enqueue(mockedDeletePipelineResponse)
 
         buildFile << """
             saagie {
                 server {
                     url = 'http://localhost:9000'
-                    login = 'fake.user'
-                    password = 'ThisPasswordIsWrong'
+                    login = 'test.user'
+                    password = 'password'
                     environment = 2
                 }
 
-                project {
-                    id = 'wrong id'
+                pipeline {
+                    id = 'bad-id'
                 }
             }
         """
-        mockWebServer.enqueue(mockedResponse)
 
         when:
         BuildResult result = gradle(taskName)
@@ -142,7 +138,7 @@ class JobListTaskTests extends Specification {
         then:
         UnexpectedBuildFailure e = thrown()
         result == null
-        e.message.contains('Something went wrong when getting project jobs: {"data":null,"errors":[{"message":"Unexpected error","extensions":null,"path":null}]}')
+        e.message.contains('Something went wrong when deleting pipeline: {"data":null,"errors":[{"message":"Unexpected error","extensions":null,"path":null}]}')
         e.getBuildResult().task(":${taskName}").outcome == FAILED
     }
 }
