+++ src/test/groovy/io/saagie/plugin/tasks/pipeline/PipelineGetInstanceStatusTaskTests.groovy
@@ -1,0 +1,112 @@
+package io.saagie.plugin.tasks.pipeline
+
+import okhttp3.mockwebserver.MockResponse
+import okhttp3.mockwebserver.MockWebServer
+import org.gradle.testkit.runner.BuildResult
+import org.gradle.testkit.runner.GradleRunner
+import org.gradle.testkit.runner.UnexpectedBuildResultException
+import org.junit.Rule
+import org.junit.rules.TemporaryFolder
+import spock.lang.Shared
+import spock.lang.Specification
+import spock.lang.Title
+
+import static org.gradle.testkit.runner.TaskOutcome.FAILED
+
+@Title('projectsGetPipelineInstanceStatus task tests')
+class PipelineGetInstanceStatusTaskTests extends Specification {
+    @Rule TemporaryFolder testProjectDir = new TemporaryFolder()
+    @Shared MockWebServer mockWebServer = new MockWebServer()
+    @Shared String taskName = 'projectsGetPipelineInstanceStatus'
+
+    File buildFile
+    File jobFile
+
+    def setupSpec() {
+        mockWebServer.start(9000)
+    }
+
+    def cleanupSpec() {
+        mockWebServer.shutdown()
+    }
+
+    def setup() {
+        buildFile = testProjectDir.newFile('build.gradle')
+        buildFile << 'plugins { id "io.saagie.gradle-saagie-dataops-plugin" }\n'
+
+        jobFile = testProjectDir.newFile('jobFile.py')
+    }
+
+    def cleanup() {
+        mockWebServer.dispatcher.peek()
+    }
+
+    private BuildResult gradle(boolean isSuccessExpected, String[] arguments = ['tasks']) {
+        arguments += '--stacktrace'
+        def runner = GradleRunner.create()
+            .withArguments(arguments)
+            .withProjectDir(testProjectDir.root)
+            .withPluginClasspath()
+            .withDebug(true)
+
+        return isSuccessExpected ? runner.build() : runner.buildAndFail();
+    }
+
+    private BuildResult gradle(String[] arguments = ['tasks']) {
+        gradle(true, arguments)
+    }
+
+    def "projectsGetPipelineInstanceStatus should return the status of the pipelineInstance"() {
+        given:
+        def mockedRunJobResponse = new MockResponse()
+        mockedRunJobResponse.responseCode = 200
+        mockedRunJobResponse.body = '''{"data":{"pipelineInstance":{"status":"SUCCEEDED"}}}'''
+        mockWebServer.enqueue(mockedRunJobResponse)
+
+        buildFile << """
+            saagie {
+                server {
+                    url = 'http://localhost:9000'
+                    login = 'fake.user'
+                    password = 'ThisPasswordIsWrong'
+                    environment = 2
+                }
+
+                pipelineinstance {
+                    id = "pipelineInstanceId"
+                }
+            }
+        """
+
+        when:
+        BuildResult result = gradle(taskName)
+
+        then:
+        !result.output.contains('"data"')
+        result.output.contains('"status"')
+    }
+
+    def "projectsGetPipelineInstanceStatus should fail if no pipelineInstance id was provided"() {
+        given:
+
+        buildFile << """
+            saagie {
+                server {
+                    url = 'http://localhost:9000'
+                    login = 'fake.user'
+                    password = 'ThisPasswordIsWrong'
+                    environment = 2
+                }
+            }
+        """
+
+        when:
+        BuildResult result = gradle(taskName)
+
+        then:
+        UnexpectedBuildResultException e = thrown()
+        result == null
+        e.message.contains("Missing params in plugin configuration: https://github.com/saagie/gradle-saagie-dataops-plugin/wiki/${taskName}")
+        e.getBuildResult().task(":${taskName}").outcome == FAILED
+    }
+}
