+++ src/test/groovy/io/saagie/plugin/tasks/platform/PlatformListTaskTests.groovy
@@ -1,4 +1,4 @@
-package io.saagie.plugin.platform
+package io.saagie.plugin.tasks.platform
 
 import io.saagie.plugin.dataops.DataOpsExtension
 import io.saagie.plugin.dataops.utils.SaagieUtils
@@ -10,8 +10,6 @@ import org.gradle.testkit.runner.GradleRunner
 import org.gradle.testkit.runner.UnexpectedBuildFailure
 import org.junit.Rule
 import org.junit.rules.TemporaryFolder
-import spock.lang.Ignore
-import spock.lang.IgnoreRest
 import spock.lang.Shared
 import spock.lang.Specification
 import spock.lang.Title
